Configure: new --enable-sanitize to help in debugging
Until now, it was not possible to easily configure Gnuastro with the
compiler sanitizer option(s): '-fsanitize'.

With this commit, a new '--enable-sanitize' option has been added to the
configure script for this and a corresponding option has been added to the
developer-build script. Both are also documented in the book.