color-faint-gray: crash using 4th channel bug fixed
Until now, this script works fine if three images are used as inputs,
but it crashes if an extra 4th channel is used. The problem is related
to the creation of a symbolic link. In fact, we commented in the source
code of the script that we should avoid using symbolic links because
'realpath' is not portable on non-GNU systems

With this commit, I have fixed this bug. I also removed other remaining
symbolic links into the code.

This fixes bug #66303 (https://savannah.gnu.org/bugs/index.php?66303)
reported by Elisabeth Sola.