color-faint-gray: option --rhdu renamed to --reghdu (regions HDU)
Until now, the HDU variable for the R channel image ('rhdu' within the
code) was confused with the HDU variable of the regions image specified
by the option '--rhdu'.

With this commit, that problem has been fixed. Now the option for
specify the HDU of the region image is '--reghdu' or '-r'. I also
noticed that this option was not described in the Book, so I added the
necessary information.

In addition to that, there was another bug caused by the use of 'cp' to
copy an input (4th channel) image without specifying its HDU properly.
Now, instead of copying with 'cp', the image is copied by using 'astfits'
with the proper '--copy' option. By doing this, the output image is in the
HDU=1 by default. Finally, Elizabeth Sola's contribution was added in
Gnuastro's relevant files.