MakeCatalog: known bug warning printed for in-slice of clumps
Until now, when in-slice measuremens were requested for clumps within
MakeCatalog, an ambiguous message would be printed about an un-recogonized
column and MakeCatalog would crash! This was because we have not yet had a
chance to implement the in-slice columns for clumps within MakeCatalog.

With this commit, those specific columns have been added for clumps, but
instead of printing the generic/ambiguous message, it now prints the
precise cause of the problem with a link for people to follow up on its
status.

This issue was reported by Sepideh Eskandarlou.