Library (arithmetic.h): new sblim-diff for SB Limit of new project
Until now, in the "Surface brightness limit of image" section of the book,
we had only defined the surface brightness limit. But given its
theoretical/extrapolated nature, it is easy to also estimate this limit for
other telescopes and exposure times.

With this commit, the necessary descriptions to calculate this are added to
that section. To simplify its derivation without needing a calculator, the
new 'sblim-diff' operator was added to Arithmetic.