developer-build: the prefix of --install is found automatically
Until now, this script would always install in the default '/usr/local'
prefix if no prefix was given. But when the user already has their
installed version of Gnuastro in another location, it was had to give the
prefix every time.

With this commit, the script will extract the prefix of the existing
Gnuastro installation and install Gnuastro there.