astscript-fits-view: new --topcat4k for high density monitors
Until now, when TOPCAT was called on 4K laptop monitors (high density
monitors), it would be too small to use! TOPCAT has two ways of dealing
with such cases, but until now, there was no way to pass them to TOPCAT
when using this script. Also, there was a typo in the example of the
recently added 'sblim-diff' operator (the effective radii were calculated
from diameters!).

With this commit, through the new '--topcat4k' option, this script will
call TOPCAT in the proper way so it is scaled by two times. Also, the typo
in the example 'sblim-diff' has been corrected.