MakeProfiles: ignoring pixels smaller than floating point minimum
Until now, MakeProfiles would build all the pixels of a profile until the
truncation radius. Even when the truncation radius is so large that the
values can be below the minimum 32-bit floating point (and thus
meaningless: the floating point error dominates).

With this commit, when the profile's pixels approach the lowest floating
point value, the truncation radius is changed so pixels beyond that radius
are no longer built (that can take a lot of time).

During the debugging process, I noticed that the fact that the program do
not print the extra argument when more than one argument is given can be
annoying. With this commit, all the relevant functions in the programs that
have this error also print the extra argument to help the user understand
and fix the problem.