MakeProfile: correct first-pixel check to avoid passing truncation
Until now, when MakeProfiles was given a small profile with a strong
gradient (such that all the pixels need monte carlo integration: which has
a random element!), in some cases, it could continue building the profile
after the truncation radius! For an example command (and corresponding
image) see the bug report: https://savannah.gnu.org/bugs/?66572

After inspecting the code, the problem was discovered to be due to the
conflict between the way we checked for being the first-pixel, and the
radius of the pixel (to be checked with the trucation radius). In short,
when the truncation is smaller than one pixel, we want to at least have
that single pixel and truncate after that. To check if this was the first
pixel, we were using the larger end of the linked list of pixels to
add. But when the gradient is strong (and monte carlo integration is
necessary for all pixels), the list can sometimes become empty, allowing
MakeProfiles to ignore the truncation radius!

With this commit, instead of a check on the list, a new binary variable was
added that immediately flips after the first pixel. With this, it is no
longer possible to ignore the truncation radius after the first pixel.