Library (blank.h): gal_blank_trim returns 0x0 if all image is blank
Until now, when the input to this function (which is used by the 'trim'
operator of Arithmetic) was all blank, it would produce an output that has
2x2=4 pixels (all still blank!)!  But this is not correct: it should be 0x0
(there are no pixels with a non-blank value to remain!).

With this commit, the problem was solved by checking the final output
coordinates and creating an output array that has no elements.

This fixes bug #66588.