Library (table.h): introduce gal_table_sort
Until now, it was possible to sort a table from the command line ('asttable
--sort ...'), but no similar high-level function was available in the
library (for use in other programs or for users of the library).

With this commit, the sorting function that was used in the option above
(and was only available within the Table program) is moved to 'table.h'.