Webpage: added image of ADS citation statistics per year
Until now, the webpage simply had a section at the end, pointing to the ADS
link that viewers had to click and navigate to see the citation
statistics. That was inconvenient and many viewers to the top page would
miss it. On a side-note, a small edit was also found to be necessary for
the example convolve commands in the book to find the matching PSF.

With this commit, the top of the webpage shows the citation chart of
Gnuastro (linking to the bottom for interested readers) and the description
and example command in the Invoking Convolve program have also become more
clear.

Both these changes were suggested by David Valls-Gabaud.