astscript-psf-scale-factor: also report the STD value as an output
Until now, only the scaling factor computed as the sigma clipping median of
the ring of pixels was reported as the output. However, there are some
situations in which it is interesting to have the standard deviation value
of such a ring o pixels. For example, if the user is trying to re-fine the
center position of the PSF with respect to the star. The best center
position will be that position (x,y) that minimizes the standard deviation
of the ring.

With this commit, minor changes have been added to report the standard
deviation value as a second number in the output of the 'psf-scale-factor'
script. Necessary information explaining this new feature has been also
included into the Book.