Book: TeXLive instal instructions no longer need root permission
Until now, the instructions to install TeXLive for the bootstrapping
dependencies of Gnuastro simply assumed the default instation directory
(which needed root permissions!). But it is generally better to avoid
becoming root for security and convenience!

With this commit, the descriptions on how to install TeXLive are more
complete and include the steps to install it in the user's home directory.