Book: specifying queried columns when checking for overlapping sources
Until now, within the code example of the '--overlapwith' option of query,
when checking for sources that overlap with the input image after a catalog
was retrived, the final output catalog contained an extra column
('NUM_INPUTS') that was not requested by the user.

With this commit, the requested columns are also specified in the last
command so the output catalog contains only the requested columns by the
user.