PSF scripts: continue if Crop does not create an output image
Until now, if Crop was not able to create an output stamp because the
specified coordinates/width did not overlape with the image, then the
execution was stopped because of the use of -e shell flag in the
scripts. This is not good in general because it is an expected situation in
large pipelines and data processing.

With this commit, an 'if !' statement has been added before the crop
command to manage the situation elegantly. When Crop creates the image,
then everything is fine. If the image is not created, then we create an
image full of NaN values (in 'astscript-psf-stamp'), and return NaN values
(in 'astscript-psf-scale-factor'). As a consequence, now, even when the
cropped image is not generated, the execution will continue without
stopping the processing.