Library (statistics.h): fixed incorrect integer type checking
Until now, the check to make sure the input to
'gal_statistics_concentration' is an integer was mistakenly using an '||'
between the two float checks. As a result, floating point inputs could have
leaked in.

With this commit, the condition is properly checked with an '&&'.

This issue was reported by Faezeh Bidjarchian.