astscript-radial-profile: undersample=1 accounted for correctly
Until now, the undersampling was performed as soon as a value for
'--undersample' was provided. However, there is one case for which it is
not necessary to perform the undersampling even if it is asked
explicitly. This situation is '--undersample=1'. This situation caused the
filling of the central pixel (0 value) two times, but it is only necessary
to do it once since undersampling by 1 means do nothing.

With this commit, the script checks that the undersampling factor is
different from 1. By doing this, only significant and non-trivial
undersampling are performed. This avoids also having the central pixel
different from 1 as a consequence of multiple re-filling steps.

This bug was reported by Samane Raji.

This fixes bug #67105.