Library (pdf.h): new check for Ghostscript existence
Until now, when the host system didn't have Ghostscript, the
'gal_pdf_write' function would crash during Gnuastro's build (in the 'make'
phase), complaining that the 'PATH_GHOSTSCRIPT' macro is not defined!

With this commit, an '#ifdef' macro check has been placed around that part
of the code (as we do with other libraries based on optional dependencies)
so no crash occurs. In such cases, the calling program will crash at
run-time with a descriptive error message, but it will atleast create the
EPS file that the users can convert to PDF by other means until they
install Ghostscript and later re-install Gnuastro.

This bug was reported by Faezeh Bidjarchian.

This fixes bug #67106.