Library (blank.h): new blank_flag_remove arg. for optional freeing
Until now, when the input to 'blank_flag_remove' was an all-blank array,
this function would free its allocated space. However, this caused a crash
in the stacking operators when all the inputs where blank.

This was because we keep a single allocated array for all the pixels of the
coadd and constantly write values inside of it. But when all the input
pixels to one output pixel were blank, 'blank_flag_remove' freed the array
that the subsequent pixels would need.

With this commit, the problem is solved by adding an extra argument
'blank_flag_remove', allowing the caller to optionally free the input or
not. All other calls to this function have been modified with this commit
to have the same behavior as they had until now (free if input was all
blank).