Library (arithmetic.h): maskfilled operators account for 0 scatter
Until now, the two 'madclip-maskfilled' and 'sigclip-maskfilled' operators
didn't account for situations where the measured spread of the
distributions was 0 (can happen in integer images like dark exposures for
example). After multiplying by the desired factor, the scatter would still
be 0, which would effectively mask all most of the pixels in each image!

With this commit, when the scatter on an output pixel is zero, we change
the upper and lower limit to 'inf' and '-inf' to keep all the elements.

This fixes bug #67128.