Library (arithmetic.h): multiple of sigma/MAD passed to maskfilled
Until now, the maskfilled operators didn't take into the account the
multiple of the scatter that the user provided. This was because in the new
implementation of the ranges of the previous commit (commit 2bdfe9052), I
had forgot include it! Furthermore:

  1. The Arithmetic '--meta*' options (to set the name, units and comments
     of the output) only worked on the top output of multiple-output
     Arithmetic calls, not the rest.

  2. 'statistics_mad_in_sorted_no_blank' and 'gal_data_copy_to_allocated'
     didn't account for a 'gal_data_t' where its 'size' was zero; causing
     segmentation faults.

With this commit, all the issues above have been addressed: the given
multiple of the scatter is now implemented in the 'maskfilled' operators,
the two library functions account for scenarios were the 'size' of a
dataset is zero and the '--meta*' options of Arithmetic accept multiple
values.