Library (statistics.h): MAD-clipping when MAD is zero
Until now, when the MAD of a distribution was zero (can happen in integer
datasets with repeated medians), MAD-clipping was effectively useless.

With this commit, a solution has been implemented where a new measure of
spread is defined in such cases: the difference between the values adjacent
to the median (fully described in the 'MAD_ALTERNATIVE' macro of
'lib/statistics.h').

Also, the following edits have been implemented as part of this commit:

  - The 'gal_statistics_minimum' and 'gal_statistics_maximum' now check if
    the input is sorted and has no blanks. If so, they will just use the
    respective index of the array to avoid having to parse the input all
    the way to the end.

  - A new 'gal_statistics_range_double' function has been added which was
    written while exploring different solutions. It was not actually used
    for the MAD clipping goal of this commit, but may nevertheless be
    useful in other scenarios, so it was kept.