Libarary (units.h): apparent magnitude to luminosity and vice-versa
Until now, we did not have any operators in the Arithmetic or Table (in
columns arithmetic) programs to convert apparent magnitude to the
luminosity (in uints of solar luminosity) and vice-versa.

With this commit, new functions are added to 'lib/unit.c' and based on
them, two new operators are added to the 'lib/arithmetic.c' libarary which
is used in the Arithmetic and Table programs. The functions take the
distance modulus for the conversion.