Library (txt.c): an empty input is given a dsize value of 0
Until now, when an empty text file was read by Gnuastro's plain-text
library, it would free the internal 'dsize' array (to keep the length of
the dataset along different dimensions). However, this caused a
segmentation fault in later steps (which by default assume that 'dsize' is
allocated, but has a value of 0).

In parallel, I noticed that the description of the '--append' option of
Arithmetic, actually had the description of the same option in the Crop
program!

Wit this commit, this has been fixed by not freeing the array in
'lib/txt.c' in such cases, but instead writing a value of zero inside of
it. The description of Arithmetic's '--append' option has also been
corrected.

The plain-text table reading bug was reported by Sepideh Eskandarlou.

This fixes bug #67190.