Arithmetic: correctly parses -inf and -nan
Until now, if Arithmetic was called like below it would complain with an
invalid '-i' option: 'astarithmetic 2 -inf x'. This happened because before
parsing for options, we checked all arguments starting with a dash and
checked if the next character is a digit or not. If so, we would treat them
differently. However, this didn't account for special numbers like '-inf'
or '-nan'!

With this option, the problem has been solved by also check for '-inf' and
'-nan'.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #67191.