Book: using the term coadd for adding many data sets into one
Until now, the term "stack" was used in two contexts within the Arithmetic
program: referring both to the stack of operands during its operation and
the coadding process (when we add many images to make a deeper one).

With this commit, when talking about the latter scenario, the book now uses
the term "Coadd" (which is also commonly used in the optical astronomical
community). One of the operators of the 'astscript-pointing-simulate'
script also needed to be changed to unify the Gnuastro approach.