library (fits.h): informative error when output FITS has parenthesis
Until now, when an output FITS file had a parenthesis in it, and it didn't
fit CFITSIO's convention for parenthesis, Gnuastro's programs would simply
abort with the cryptic CFITSIO error. This was very hard to understand for
a regular user.

With this commit, when CFITSIO fails to open a FITS file for writing, we
now check for parenthesis in it, and if there is a parenthesis, an
informative text is printed to help the user easily address the problem.

This issue was found with the help of Sepideh Eskandarlou.