Library (units.h): mag-to-luminosity takes the Sun's absolute mag
Until now, this newly added operator would assume the sun's bolometric
absolute magnitude, which is only useful if you have the bolometric
apparent magnitude! But in most cases, we have the apparent magnitude in a
certain filter. So the user should be able to set the sun's absolute
magnitude in the same filter.

With this commit, the units.c functions and Arithmetic operators take three
operands, not two (which now includes the user-provided absolute magnitdue
of the Sun). A reference paper has also been added in the book to help
users find the proper value.