Book: elaborated the description of the distance modulus and luminosity
Until now, the distance modulus and '--absmagconv' options didn't have any
useful information: only a single setence; that was not too clear on the
purpose of their difference; and thus which one should be used for what
scenario.

With this commit, a complete description has been added for the issues with
the distance modulus and the related parts of the book (including the
brightness, flux and magnitude section) have been edited to better
integrate this definition.

In a separate situation, I also noticed that in the Quick start section
didn't link to the "Installation directory" section to remind the users
that being root is not necessary, so a sentence was also addded
there. Furthermore, the newly added units functions weren't mentioned in
the NEWS file, so they were also added.