Arithmetic: --append will not write any value in the 0th HDU
Until now, when Arithmetic was called with the '--append' option, it would
continue to write its metadata (all the option values given to it) in the
0-th HDU. However, when appending to the output, we just want to add new
HDUs, not edit the existing ones.

With this commit, with the '--append' option, Arithmetic will not update
the 0th HDU of the already existing file.