Mkcatalog: introduce Confusion Limit in headers
Until now, Confusion Limit statistics were not present in the FITS
headers of Mkcatalog's output.

With this commit, if a clumps catalog is given, the program computes the
(MAD-clipped) median distance in pixel between a clump and its nearest
neighbour, and converts it in arcseconds if also a wcs is provided. MAD and
standard deviation of the distance are also added to the header.

Incidentally, the nearest neighbour search function of the kdtree.h library
and the k-d tree based matching function of match.h are updated to ignore
exact matches if the (newly introduced) "nosamenode" argument is not zero.