Table: --equal can be called on same col with diff values separately
Until now, when Table's '--equal' operator was called multiple times on the
same column with different values the output will be an empty table (the
first '--equal' will only keep a certain set of rows that do not match the
second '--equal'!). This would not happen when the values were given in one
call to '--equal'. This is contrary to the description of the '--equal' (or
'--notequal') in the book (which explicitly says that both modes are fine).

With this commit, the problem was solved by redefining an alternative to
the default parser of '--equal' or '--notequal'. An internal appending
function also became necessary in 'data.h'.