MakeCatalog: reorganized output keywords and metameasurements
Until now, MakeCatalog calculated the measured surface brightness limit
(SBL, a metameasure) of the input image automatically when the standard
deviation image was present. With the previous commit, MakeCatalog also
calculated the confusion limit (another metameasure). But it is not as
computationally cheap as the SBL and should not be executed all the
time. In general, MakeCatalog's code and documentation were not really
friendly for all the metameasures that are generally necessary for a good
data analysis.

With this commit, MakeCatalog's final keyword writing parts where
re-organized into a new 'outkeys.c' file to help in the organization and
future expansion of metameasures. All of this has been implemented as the
following changes:

  - Metadata (including metameasures) are written the 0th HDU.
  - To have metameasures, the user needs to call the '--meta-measure'.
  - The keywords have been standardized and better organized.
    - The old SBL keywords have also changed.
  - The noise-based magnitude limit has also been added.
  - The sections of MakeCatalog in the book have undergone a major
    restructuring.

This commit and the previous one were based on the good suggestion by
Rafael Guzman.