Library (tile.h): new function to put a tile over a labeled dataset
Until now, Gnuastro had a function to define a tile over each label of the
input, but it was within the MakeCatalog program and not accessible within
the library (to be used in other scenarios).

With this commit, the function has been extracted out of MakeCatalog and
brought into the library as 'gal_tile_per_label'.