Library (label.h and tile.h): new functions to measure over labels
Until now, only MakeCatalog contained a function to define tiles over a
label and parse over the label and a values data set to make some
measurements. However, this feature is useful outside of MakeCatalog also!
In particular within Arithmetic when we need to select pixels associated
with a label that has certain features.

With this commit, the function that MakeCatalog used has now been
generalized and moved into Gnuastro's 'tile.h' library and a new function
has been added to the 'label.h' library to use it within the Arithmetic
library and thus program.

In the process, I also noticed a bug in MakeCatalog that would occur when
the input to MakeCatalog was not from a raw Segment output with clumps (for
example if the input was a crop of Segment) and have fixed it.

This fixes bug #67275.