Fits: output name only set after input files are read
Until now, the Fits program would crash with a segmentation fault if no
input file was specified and a HDU relation operation (that needed an
output name) was requested. This was because we would try to set the output
name for those operations before checking if any input was provided.

With this commit, the part of setting the output name has been moved after
we have verified that an input has been given. Therefore, Fits now aborts
elegantly (with an informative message to help the user find the problem in
their scripts).