MakeCatalog: arrays for clump labs when input not from Segment
Until now, MakeCatalog used a linked list for finding clump unique clumps
labels in the input when it was not created by Segment. Since this involves
parsing through the list for every clump pixel, it could become extremely
inefficient for deep images (where almost all pixels are
clumps!). Furthermore, the ordering of the clumps in the final catalog
would be based on the first clump that is found in the image, not their
actual labels.

With this commit, to fix both problems, we are now using an array instead
of a list. To match the labels we can now simply use the index within the
array (which is very fast) instead of parsing the list (which is extremely
slow) and also the clumps table in the output is sorted based on the clump
ID automatically.