Library: match functions flag ambiguous matches
Until now, when one point of an input catalog had multiple corresponding
(within the aperture) points in the other catalog Match program's outputs
(for inner or full arrangements) would be ambiguous and would depend on the
order that the catalogs were given.

With this commit, the low-level functions of sort-based and k-d tree based
matching have been corrected to flag such cases and the Match program also
now doesn't include them in its main outputs and writes a new
'FLAGGED-IN-2ND' HDU to inform their user of the cases.

However, an important point about the k-d tree solution is that it is buggy
(will not catch all the objects within the given aperture); For more see
https://savannah.gnu.org/bugs/index.php?67364

In the middle of work on this, the following points were also implemented:

  - Match's default algorithm has is now sort-based due to the bug (will be
    switched back to k-d tree after the bug is fixed).

  - Match's output will always be a FITS file from now on.

  - The '--outcols' option can be called multiple times in one command.