Statistics: fitting outputs printed in scientific notation
Until now, the fitting outputs of the Statistics program on the
command-line used printf's '%f' format, so for very small numbers, it would
only show zeros (see bug report).

With this commit, they have been converted to '%e' to use scientific
notation and fix this problem.

This bug was reported by Faezeh Bidjarchian.

This fixes bug 67475.