ds9-region: new --fontsize option
Until now, it was not possible to easily change the font of the name-column
for DS9 regions automatically. It was necessary to edit the region file
manually. Another DS9 related issue was that DS9 recently has defaulted to
ICRS instead of FK5, causing issues when creating polygons in DS9 and using
the default value.

With this commit a new '--fontsize' option has been added to the
'ds9-region' script. Also, 'lib/ds9.c' has been modified to read both 'fk5'
and 'icrs' as WCS coordinates and the 'fits-view' script now also defaults
to ICRS (which is more recent than FK5).