MakeCatalog: message when clumps and objects from different images
Until now, in special situations where the clumps and objects were derived
from different images, MakeCatalog would either crash with a segmentation
fault or keep running without stop.

With this commit, the problematic point has been detected. When this
situation occurs, an elaborate error message is printed (guiding the user
on the cause of the problem and what to do afterwards) and MakeCatalog
aborts.