Book: --write=/ in the Fits program points to the FITS standard
Until now, the description of this mode of the '--write' option in the Fits
program didn't use a matching terminology to the FITS standard (which
treats such lines as "commentary keywords").

With this commit, the terminology has been fixed and a paragraph has been
added to directly describe this behaviour while linking to the relevant
sections of the FITS standard.