Library (fit.h): support 2d polynomial fitting
Until now, polynomial fitting was performed using a 1d polynomial.

With this commit, 2d polynomials are supported. The function names
have changed to 'gal_fit_polynomial' and 'gal_fit_polynomial_robust'.
The code select automatically between the two polynomials depending on
the number of elements in the list given as first argument.