WIP: interface to 2D fitting in Statistics
Until now, the low-level fitting function was added in the commits before
this, but there was no direct/easy command-line interface to test/use them.

With this commit, the low-level functions have been integrated into the
Statistics program (except for its estimation features; due to lack of
time). The fitting library function names and code have also been
generalized to have the number of dimensions at the end.