Book and NEWS: corrected mistakes in previous commit
Until now (in the previous commit), I had forgot to add one of the newly
added library functions, mistakenly kept the edits in 'tests/during-dev.sh'
and added a vaguely termed box about the estimation of the fit within the
documentation. To make things worse, I had also forgot to remove the "WIP"
(work in progress) from the commit message title.

With this commit, the ones that could be edited (except for the commit
message title) have been corrected.