zeropoint: missing options on the command-line added
Until now, the '--starcatra' and '--starcatdec' options were defined in the
script, but they were missing from the "Separate command-line arguments
from options" section of the zeropoint script. Because of this, when using
these options on the command line, the script failed with the error:
"unknown option '--starcatra=ra'".

With this commit, the missing options have been added to the proper
section. In addition, a clear explanation on how to set the 'utilname' for
scripts have been provided in the 'tests/during-dev' script.