Library (fit.h): avoiding uninitialized variable warning
Until now, the 'fit_polynomial_nconst' function of 'lib/fit.c' didn't
initialize its main variable and also didn't crash with an error if the
'matrixid' was not recognized (in case of a bug!).

With this commit, both issues have been addressed. I also noticed that the
indentation was not standard, so that was also fixed.