Book: improve example regarding *clip-all
Until now, the example for the '*clip-all' operators suggested to create a
temporary file with all the generated information (mean, std, median, mad,
number), and then invoke 'astfits' to copy only the desired HDUs. This
approach is not optimal and would affect performance on systems where the
number of i/o operations is relevant (such as a pipeline running on a large
number of threads in a HPC system with network storage).

With this commit, it is instead suggested to use a combination of 'set-'
and 'free' inside the astarithmetic command, in order to save in the output
file only the required info, without the need for a temporary file.