Match: fix typo in warning bug in k-d tree match
Until now, the condition to warn the user of the issue that multiple
matches within the aperture might not be flagged and that results might be
dependent on the order of inputs was always false due to a typo (an "and"
instead of an "or").

With this commit, the condition is fixed.