Library (match.h): initialization to avoid warning
Until now, the 'ai' variable in the 'match_kdtree_worker' function could
produce a "maybe used uninitialized" warning during compilation. Also, in
the prevoius commit, I forgot to add the bug fix in the NEWS file.

With this commit it is initialized to an impossible value to avoid this
warning and also cause a crash if it is ever used without being properly
set. Also, the bug fix is now mentioned in the NEWS file.