Match: default value of --kdtree set to internal
Until now, the default k-d tree strategy of the Match program was for it to
be disabled bug #67364. However, that bug was fixed a few commits ago
(commit 71d09773), so it is inefficient for k-d tree to stay disabled by
default.

With this commit, the default k-d tree strategy is set to 'internal'.