Match: full arrangement and no match merge the full catalogs
Until now, when the Match program was asked to do a full arrangement with
'--outcols' activated, it would produce an empty table! This is not
expected in a full arrangement (where completely non-matching catalogs can
be present and the user expects them to be merged).

With this commit, the necessary corrections have been made for the full
match to behave as expected.